整理代码
对验证性实验的代码进行整理
